+++ lib/src/paper_world/paper_world_widget.dart
@@ -11,7 +11,8 @@ import '../world_asset/world_asset_internal/world_asset_internal.dart';
 import '../world_asset/world_asset_internal/world_asset_render_model.dart';
 import 'controller/paper_world_bloc.dart';
 
-class PaperWorldWidget extends StatefulWidget {
+class PaperWorldWidget extends StatelessWidget {
+
   final List<WorldAsset> assets;
   final BehaviorSubject<CameraModel> camera;
   final BehaviorSubject<Size> screen;
@@ -20,50 +21,36 @@ class PaperWorldWidget extends StatefulWidget {
   const PaperWorldWidget({Key? key, required this.assets, required this.camera, this.background, required this.screen}) : super(key: key);
 
   @override
-  State<PaperWorldWidget> createState() => _PaperWorldWidgetState();
-}
-
-class _PaperWorldWidgetState extends State<PaperWorldWidget> {
-
-  @override
-  void dispose() {
-    for (var asset in widget.assets) {
-      asset.dispose();
-    }
-    super.dispose();
-  }
-
-  @override
   Widget build(BuildContext context) {
-    return Container(
-      color: widget.background ?? Colors.white,
-      height: MediaQuery.of(context).size.height,
-      width: MediaQuery.of(context).size.width,
-      child: MultiProvider(
-        providers: [
-          BlocProvider<PaperWorldBloc>(create: (context) => PaperWorldBloc(widget.assets)),
-          Provider<BehaviorSubject<CameraModel>>.value(value: widget.camera)
-        ],
-        child: BlocBuilder<PaperWorldBloc, GameWorldState>(
-          builder: (context, state) => state is GameWorldReady ?
-            Stack(
+    return MultiProvider(
+      providers: [
+        BlocProvider<PaperWorldBloc>(create: (context) => PaperWorldBloc(assets)),
+        Provider<BehaviorSubject<CameraModel>>.value(value: camera)
+      ],
+      child: BlocBuilder<PaperWorldBloc, GameWorldState>(
+        builder: (context, state) => state is GameWorldReady
+          ? Container(
+            color: background ?? Colors.white,
+            height: MediaQuery.of(context).size.height,
+            width: MediaQuery.of(context).size.width,
+            child: Stack(
               children: state.assetIds.map((assetId) => MultiProvider(
                 key: Key(assetId),
                 providers: [
-                  Provider<WorldAsset>.value(value: widget.assets.firstWhere((a) => a.id == assetId)),
+                  Provider<WorldAsset>.value(value: state.allAssets.firstWhere((a) => a.id == assetId)),
                   Provider<BehaviorSubject<WorldAssetRenderModel>>(
                     create: (context) => combineSubject3<WorldAssetModel, CameraModel, Size, WorldAssetRenderModel>(
                       context.read<WorldAsset>().controller.animation,
-                      widget.camera,
-                      widget.screen,
+                      camera,
+                      screen,
                       (model, camera, screenSize) => WorldAssetRenderModel(camera, model, screenSize)
                     )
                   )
                 ],
                 child: const WorldAssetInternal()
               )).toList()
-            ) : SizedBox()
-        )
+            )
+          ) : SizedBox()
       )
     );
   }