+++ lib/src/util/blocToSubject.dart
@@ -3,3 +3,6 @@ import 'package:rxdart/subjects.dart';
 
 BehaviorSubject<T> blocToSubject<T>(Bloc<dynamic, T> bloc) =>
     BehaviorSubject<T>.seeded(bloc.state)..addStream(bloc.stream);
+
+BehaviorSubject<T> cubitToSubject<T>(Cubit<T> cubit) =>
+    BehaviorSubject<T>.seeded(cubit.state)..addStream(cubit.stream);